fix: move log level globals to separate log_utils.c for link visibility
g_log_level, fp1, and log_utils_init() were defined in utils.c, which
is only compiled into libvgpu.so. The shrreg-tool executable links
multiprocess_mod (which uses LOG_* macros referencing g_log_level)
but does not link utils.o, causing undefined reference errors.

Fix: extract these definitions into a standalone log_utils.c and add
it to both the main libvgpu library and the shrreg-tool executable
in the CMake build files.
