+++ src/multiprocess/CMakeLists.txt
@@ -3,6 +3,6 @@ add_library(multiprocess_mod OBJECT multiprocess_memory_limit.c multiprocess_uti
 target_compile_options(multiprocess_mod PUBLIC ${LIBRARY_COMPILE_FLAGS})
 target_link_libraries(multiprocess_mod PUBLIC nvidia-ml)
 
-add_executable(shrreg-tool shrreg_tool.c)
+add_executable(shrreg-tool shrreg_tool.c ${CMAKE_CURRENT_SOURCE_DIR}/../log_utils.c)
 target_link_libraries(shrreg-tool multiprocess_mod -lpthread -lcuda)
